Write form values to firestore
React provides the form values as JSON to the firebase cloud function, which then stores the values to firestore.
Upon successful write, run the apps-script and provide the form values as a parameter.
