+++ .gitignore
@@ -1,6 +1,6 @@
-/node_modules/*
-/build/*
+*/node_modules/*
+*/build/*
 
-/.env
+*/.env
 *credentials.json
 *service-key.json
