+++ apps-script/Code.js
@@ -1,0 +1,6 @@
+function myFunction(data) {
+  
+  var party = JSON.parse(data)
+  console.log(party)
+  return "It worked!"
+}
