+++ firebase/functions/package.json
@@ -12,7 +12,7 @@
     "node": "8"
   },
   "dependencies": {
-    "firebase-admin": "^8.0.0",
+    "firebase-admin": "^8.6.1",
     "firebase-functions": "^3.0.0",
     "googleapis": "^41.0.1"
   },
