+++ firebase/package.json
@@ -3,18 +3,18 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@date-io/date-fns": "latest",
-    "@material-ui/core": "^4.1.2",
-    "date-fns": "next",
-    "final-form": "^4.16.1",
+    "@date-io/date-fns": "^1.3.11",
+    "@material-ui/core": "^4.5.1",
+    "date-fns": "^2.0.0-beta.5",
+    "final-form": "^4.18.5",
     "final-form-material-ui": "^0.3.0",
-    "firebase": "^6.2.2",
+    "firebase": "^6.6.2",
     "googleapis": "^39.2.0",
     "material-ui-pickers": "^2.2.4",
-    "react": "^16.8.6",
-    "react-dom": "^16.8.6",
+    "react": "^16.10.2",
+    "react-dom": "^16.10.2",
     "react-final-form": "^6.3.0",
-    "react-router-dom": "^5.0.1",
+    "react-router-dom": "^5.1.2",
     "react-scripts": "3.0.1",
     "recompose": "^0.30.0"
   },
