+++ firebase/public/manifest.json
