--- functions/index.js
@@ -1,68 +1,0 @@
-const functions = require('firebase-functions');
-const {
-  google
-} = require('googleapis');
-
-const googleCredentials = require('./credentials.json')
-
-// Create and Deploy Your First Cloud Functions
-// https://firebase.google.com/docs/functions/write-firebase-functions
-
-exports.helloWorld = functions.https.onCall((data, context) => {
-  console.log(data)
-  console.log(context)
-
-  const scriptId = '1nvPPH76NCCZfMYNWObohW4FmW-NjLWgtHk-WdBYh2McYSXnJlV5HTf42'
-  const script = google.script('v1')
-
-  const oAuth2Client = new google.auth.OAuth2(
-    googleCredentials.web.client_id,
-    googleCredentials.web.client_secret,
-    googleCredentials.web.redirect_uris[0]
-  )
-
-  oAuth2Client.setCredentials({
-    refresh_token: googleCredentials.refresh_token
-  })
-
-  return new Promise((resolve, reject) => {
-    script.scripts.run({
-      auth: oAuth2Client,
-      resource: {
-        function: 'myFunction',
-        devMode: true
-      },
-      scriptId: scriptId
-    }, (err, resp) => {
-      if (err) {
-        // The API encountered a problem before the script started executing
-        console.log('The API returned an error: ' + err);
-        reject(err)
-      }
-      if (resp.error) {
-        // The API executed, but the script returned an error.
-
-        // Extract the first (and only) set of error details. The values of this
-        // object are the script's 'errorMessage' and 'errorType', and an array
-        // of stack trace elements.
-        const error = resp.error.details[0];
-        console.log('Script error message: ' + error.errorMessage);
-        console.log('Script error stacktrace:');
-
-        if (error.scriptStackTraceElements) {
-          // There may not be a stacktrace if the script didn't start executing.
-          for (let i = 0; i < error.scriptStackTraceElements.length; i++) {
-            const trace = error.scriptStackTraceElements[i];
-            console.log('\t%s: %s', trace.function, trace.lineNumber);
-          }
-        }
-        reject(resp.err)
-      } else {
-        console.log(`Keys: ${Object.keys(resp)}`)
-        console.log("test!")
-        console.log(`RESULT: ${JSON.stringify(resp.data)}`)
-        resolve(JSON.stringify(resp.data))
-      }
-    })
-  })
-})
